Restructure SideBar component (#94)
